pyproject.toml
[project]
name = "arryn-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.7.1",
    "python-dotenv>=1.2.2",
]
